+++ uwadv/source/screens/IngameControls.cpp
@@ -133,8 +133,8 @@ void IngameRuneshelf::UpdateRuneshelf()
    Underworld::Runeshelf& runeshelf = m_parent->GetGameInterface().GetUnderworld().
       GetPlayer().GetRuneshelf();
 
-   size_t max = runeshelf.GetNumRunes() % 3;
-   for (size_t i = 0; i < max; i++)
+   size_t max = static_cast<unsigned int>(runeshelf.GetNumRunes() % 3);
+   for (unsigned int i = 0; i < max; i++)
    {
       Underworld::RuneType rune = static_cast<Underworld::RuneType>(runeshelf.GetRune(i) % 24);
