+++ uwadv/source/screens/Panel.cpp
@@ -458,7 +458,7 @@ void Panel::UpdateInventory()
    }
 
    // paste inventory slots
-   for (size_t i = 0; i < 8; i++)
+   for (unsigned int i = 0; i < 8; i++)
    {
       if (start + i < inventory.GetNumSlots())
       {
